Make scraper print html for U.S. portion of sites
The scraper will print out an ugly spit-up of html for the portion
of the craigslist sites site that contains US states and
territories. The html includes state names, cities, + urls.

This output can be piped into a new file, or into a regular
expression matcher, which should be able to pull out all of
the urls, along with city names. HOWEVER this is a sloppy fix-
more proper method would not print and need pipe, would just
save as string- but python3 throws errors when non-string
objects get tossed in with the strings by beautiful soup.