Update README with new requirements, no todo
As the repo is going to be a collaborative space, the to-do was
removed. Future issues should use githubs bug reporting probably.
Also dependencies were changed to match new requirements.txt,
and a reference was added to the fact that this is a school
project and not some half-baked personal idea.